+++ App.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import TreeInfoScreen from '@/screens/TreeInfoScreen/TreeInfoScreen';
-import { RootStackParamList } from '@/types';
+import TreeInfoScreen from '@/screens/TreeInfo/TreeInfo';
+import { RootStackParamList } from '@/types/navigation';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
