+++ src/styles/colors.ts
