+++ src/types/navigation.ts
