+++ src/types/trees.ts
@@ -1,5 +1,3 @@
-/* eslint-disable */
-
 export enum TreeStatus {
   Available = 'Available',
   StreetReady = 'Street-Ready',
@@ -16,5 +14,3 @@ export enum TreeOwnership {
   ResidentReserved = 'Reserved for Resident',
   CommunityPlantingReserved = 'Reserved for Community Planting',
 }
-
-/* eslint-enable */
